% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepreg_221210.R
\name{best.preds}
\alias{best.preds}
\title{Title Get the best models for the steps of a stepwise fit}
\usage{
best.preds(modsum, risklist)
}
\arguments{
\item{modsum}{Model summmary}

\item{risklist}{Riskset list}
}
\value{
best predictors at each step of a stepwise regerssion
}
\description{
Title Get the best models for the steps of a stepwise fit
}
