% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_cv_231230.R
\name{bsint}
\alias{bsint}
\title{Construct the bias terms for going from model layer to layer to carry forward 
an offset to mimic a linear model}
\usage{
bsint(tnsr, lasso = 0, rreturn = 1)
}
\arguments{
\item{tnsr}{an input tensor which is to be modified to mimic the linear term of a 
generalized linear model, e.g a Cox or logistic regression model}

\item{lasso}{1 if the first column is the linear estimate from a linear model, 
often a lasso model}

\item{rreturn}{1 (default) to return an R (numeric) vector, 0 to return a torch tensor}
}
\value{
a weight matrix in tensor format
}
\description{
Construct the bias terms for going from model layer to layer to carry forward 
an offset to mimic a linear model
}
