% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_cv_240422.R
\name{calceloss}
\alias{calceloss}
\title{calculate cross-entry for multinomial outcomes}
\usage{
calceloss(xx, yy)
}
\arguments{
\item{xx}{the sigmoid of the link, i.e, the estimated probabilities, i.e. xx = 1/(1+exp(-xb))}

\item{yy}{the observed data as 0's and 1's}
}
\value{
the cross-entropy on a per observation basis
}
\description{
calculate cross-entry for multinomial outcomes
}
