% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_240423.R
\name{devrat_}
\alias{devrat_}
\title{Calculate deviance ratios for CV based}
\usage{
devrat_(m2.ll.mod, m2.ll.null, m2.ll.sat, n__)
}
\arguments{
\item{m2.ll.mod}{-2 Log Likelihoods calculated on the test data}

\item{m2.ll.null}{-2 Log Likelihoods for the null models}

\item{m2.ll.sat}{-2 Log Likelihoods for teh saturated models}

\item{n__}{sample zize for the indivual foles, or number of events for the Cox model}
}
\value{
a list with devrat.cv for the deviance ratios for the indivual folds, 
and devrat, a single collective deviance ratio
}
\description{
Calculate deviance ratios for individual folds and collectively.  Calculations
are based upon the average -2 Log Likelihoods calculated on each leave out test fold
data for the models trained on the other (K-1) folds.
}
\seealso{
\code{\link{nested.glmnetr}}
}
