% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmnetr_230508.R
\name{diff_time}
\alias{diff_time}
\title{Output to console the elapsed and split times}
\usage{
diff_time(time_start = NULL, time_last = NULL)
}
\arguments{
\item{time_start}{beginning time for printing elapsed time}

\item{time_last}{last time for calculating split time}
}
\value{
Time of program invocation
}
\description{
Output to console the elapsed and split times
}
\examples{
time_start = diff_time()
time_last = diff_time(time_start)
time_last = diff_time(time_start,time_last)
time_last = diff_time(time_start,time_last)

}
