% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.glmnetr_221210.R
\name{difftime1}
\alias{difftime1}
\title{Get elapsed time in c(hour, minute, secs)}
\usage{
difftime1(time1, time2)
}
\arguments{
\item{time1}{start time}

\item{time2}{stop time}
}
\value{
Returns a vector of elapsed time in (hour, minute, secs)
}
\description{
Get elapsed time in c(hour, minute, secs)
}
