% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.glmnetr_221210.R
\name{difftime2}
\alias{difftime2}
\title{Output elapsed time and split split}
\usage{
difftime2(time_start = NULL, time_last = NULL)
}
\arguments{
\item{time_start}{beginning time for printing elapsed time}

\item{time_last}{last time for printing time split}
}
\value{
Time of call for input, as well as time-start and time-last
}
\description{
Output elapsed time and split split
}
\examples{
time_start = difftime2()
time_last = difftime2(time_start)
time_last = difftime2(time_start,time_last)
time_last = difftime2(time_start,time_last)

}
