% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_cv_230724.R
\name{dtstndrz}
\alias{dtstndrz}
\title{Standardize a data set}
\usage{
dtstndrz(datain, lasso = 0)
}
\arguments{
\item{datain}{The data matrix set to be standardized}

\item{lasso}{1 to not standardize the first column, 0 (default) to not}
}
\value{
a standarized data matrix
}
\description{
Standardize a data set
}
