% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_240423.R
\name{factor.foldid}
\alias{factor.foldid}
\title{Generate foldid's by factor levels}
\usage{
factor.foldid(event, fold_n = 10)
}
\arguments{
\item{event}{the outcome variable in a vector identifying the different potential 
levels of the outcome}

\item{fold_n}{the numbe of folds to be constructed}
}
\value{
foldid's in a vector the same length as event
}
\description{
Generate foldid's by factor levels
}
\seealso{
\code{\link{get.foldid}} , \code{\link{nested.glmnetr}}
}
