% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmnetr_240202.R
\name{get.foldid}
\alias{get.foldid}
\title{Get foldid's with branchning for cox, binomial and gaussian models}
\usage{
get.foldid(y_, event, family, folds_n, stratified = 1)
}
\arguments{
\item{y_}{see help for cv.glmnetr() or nested.glmnetr()}

\item{event}{y_ see help for cv.glmnetr() or nested.glmnetr()}

\item{family}{y_ see help for cv.glmnetr() or nested.glmnetr()}

\item{folds_n}{y_ see help for cv.glmnetr() or nested.glmnetr()}

\item{stratified}{y_ see help for cv.glmnetr() or nested.glmnetr()}
}
\value{
A numeric vector with foldid's for use in a cross validation
}
\description{
Get foldid's with branchning for cox, binomial and gaussian models
}
\seealso{
\code{\link{cv.glmnetr}} , \code{\link{nested.glmnetr}}  , \code{\link{factor.foldid}}
}
\author{
Walter Kremers (kremers.walter@mayo.edu)
}
