% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_240822.R
\name{get.id.foldid}
\alias{get.id.foldid}
\title{Get foldid's when id variable is used to identify groups of dependent 
sampling units. With branching for cox, binomial and gaussian models}
\usage{
get.id.foldid(y_, event, id, family, folds_n, stratified)
}
\arguments{
\item{y_}{see help for cv.glmnetr() or nested.glmnetr()}

\item{event}{see help for cv.glmnetr() or nested.glmnetr()}

\item{id}{see help for nested.glmnetr()}

\item{family}{see help for cv.glmnetr() or nested.glmnetr()}

\item{folds_n}{see help for cv.glmnetr() or nested.glmnetr()}

\item{stratified}{see help for cv.glmnetr() or nested.glmnetr()}
}
\value{
A numeric vector with foldid's for use in a cross validation
}
\description{
Get foldid's when id variable is used to identify groups of dependent 
sampling units. With branching for cox, binomial and gaussian models
}
\seealso{
\code{\link{factor.foldid}} , \code{\link{nested.glmnetr}}
}
