% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.glmnetr_230218.R
\name{getlamgam}
\alias{getlamgam}
\title{get numerical values for lam and gam}
\usage{
getlamgam(object, lam, gam, comment)
}
\arguments{
\item{object}{glmnetr object as input}

\item{lam}{value for lam, may be NULL, typically NULL.}

\item{gam}{value for gam, may be MULL, typically NULL, 0 or 1.}

\item{comment}{Default of TRUE to write to console information on lam and gam selected for output.
FALSE will suppress this write to console.}
}
\value{
numerical values for lam and gam for use in plot(), predict() and summary() functions
}
\description{
This functin derives the numerical values for lam and gam (lambda and gamma) for usage in plot and 
predict() functions.  If the input variables lam and gam are unspecified then 
the cross validation informed lambda and gamma values ('lambda.min' and 'gamma.mim) 
which minimize the cross validation deviance, are returned.  One may also give
lam='lambda.1se' and gam='gamma.1se' to identify the corresponding numerical 
values.  Importantly one may also simply specify gam=1 (and lam=NULL) to get the best lasso fit 
from the unrelaxed lasso model (gamma=1), or gam=0 (and lam=NULL) to get the best lasso fit
from the fully relaxed lasso model (gamma=0).
}
