% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetr_CIs_240410.R
\name{glmnetr.cis0}
\alias{glmnetr.cis0}
\title{Calculate performance measure CI's and p's}
\usage{
glmnetr.cis0(a, mu, digits = 4, txt = 0, pow = 1)
}
\arguments{
\item{a}{One term}

\item{mu}{null value}

\item{digits}{digits for printing of z-scores, p-values, etc. with default of 4}

\item{txt}{1 (default) to include inline text for estimated, 95 percent CI and p}

\item{pow}{Power to which the average of correlations is to be raised.  Only 
applies to the "gaussian" model.  Default is 2 to yield R-square but can be 1 to 
show correlations.  Pow is ignored for the family of "cox" and "binomial".}
}
\value{
An estimate, 95% CI and p
}
\description{
Calculate performances as called from glmnetr.cis().
}
