% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmnetr_221210.R
\name{glmnetr_devratio}
\alias{glmnetr_devratio}
\title{Get Deviance ratio.}
\usage{
glmnetr_devratio(object, object2, xs_new, start_new, y_new, event_new, family)
}
\arguments{
\item{object}{- A _glmnet_ output object with relax=FALSE, i.e model fit for gamma=1.}

\item{object2}{- A _glmnetr_ output object with relaxed fits, i.e model fit for gamma=0.}

\item{xs_new}{- predictor matrix}

\item{start_new}{- start times in case of usage in Cox model.  May be NULL.}

\item{y_new}{- outcome vector.}

\item{event_new}{- event indicator in case of Cox model.}

\item{family}{- Model family, one of "cox", "gaussian" or "binomial".}
}
\value{
- Deviance ratios.
}
\description{
fit models to derive the devaince ratios.
}
