% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_240301.R
\name{glmnetr_seed}
\alias{glmnetr_seed}
\title{Get seeds to store, facilitating replicable results}
\usage{
glmnetr_seed(seed, folds_n = 10, folds_ann_n = NULL)
}
\arguments{
\item{seed}{The intput seed as a start, NULL, a vector of lenght 1 or 2, or 
a list with vectors of lenght 1 or the number of folds, $seedr for most models 
and $seedt for the ANN fits}

\item{folds_n}{The number of folds in general}

\item{folds_ann_n}{The number of folds for the ANN fits}
}
\value{
seed(s) in a list format for input to subsequent runs
}
\description{
Get seeds to store, facilitating replicable results
}
