% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmnetr_230218.R
\name{glmnetrll_1fold}
\alias{glmnetrll_1fold}
\title{Evaluate fit of leave out fold}
\usage{
glmnetrll_1fold(
  object,
  xs_new,
  start_new,
  y_new,
  event_new,
  family = "cox",
  lambda_n = NULL,
  gamma = c(0, 0.25, 0.5, 0.75, 1),
  ties = "efron"
)
}
\arguments{
\item{object}{an output object from _cv.glmnetr_}

\item{xs_new}{A new predictor matrix}

\item{start_new}{A new vector of start times or the Cox model.  May be NULL.}

\item{y_new}{a new outcome vector.}

\item{event_new}{event vector in case of the Cox model.  May be NULL for other models.}

\item{family}{Model family, one of "cox", "gaussian" or "binomial".}

\item{lambda_n}{length of the lambda vector.}

\item{gamma}{The gamma vector.}

\item{ties}{method for handling ties in Cox model for relaxed model component.  Default 
is "efron", optionally "breslow".  For penalized fits "breslow" is 
always used as in the 'glmnet' package.}
}
\value{
Returns the log likelihood of object fit using new data.
}
\description{
Derive the log likelihood for a leave out based upon the fit
of the input object.
}
