% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calplot_240420.R
\name{myaxis}
\alias{myaxis}
\title{Un-log the log(HR)'s for plotting}
\usage{
myaxis(zlim, plothr, est = NULL, side = 2)
}
\arguments{
\item{zlim}{the y limits on the log(HR) scale A nested.glmnetr() output object for calibration}

\item{plothr}{a power of determining the placement of y axis 
values, e.g. 2, sqrt(10) or 10.  The default of 0 plots the X*Beta.}

\item{est}{values of X*Beta estimates to be plotted which will determine the 
range to plot in case zlim is NULL.}

\item{side}{1 for horizontal axis, 2 for vertical}
}
\value{
An axis statement
}
\description{
Un-log the log(HR)'s for plotting
}
