% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calplot_240420.R
\name{myrug}
\alias{myrug}
\title{A customized rug}
\usage{
myrug(x_, y_ = NULL, family = "gaussian")
}
\arguments{
\item{x_}{the x variable to be included in the rug}

\item{y_}{an event variable for placing the x_ values top for events and bottom
for non-events}

\item{family}{one of "cox", "binomial" or "gaussian"}
}
\value{
A rug to plot
}
\description{
A customized rug
}
