% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.stepreg_230218.R
\name{predict.cv.stepreg}
\alias{predict.cv.stepreg}
\title{Beta's or predicteds based upon a cv.stepreg() output object.}
\usage{
\method{predict}{cv.stepreg}(object, xs = NULL, ...)
}
\arguments{
\item{object}{cv.stepreg() output object}

\item{xs}{dataset for predictions.  Must have the same columns as the input predictor matrix in the call to cv.stepreg().}

\item{...}{pass through parameters}
}
\value{
a matrix of beta's or predicteds
}
\description{
Give predicteds or Beta's based upon a cv.stepreg() output object. If an input data matrix is 
specified the X*Beta's are output.  If an input data matrix is not specified then 
the Beta's are output.  In the first column values are given based upon df as a tuning 
parameter and in the second column values based upon p as a tuning parameter.
}
