% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_240322.R
\name{predict_nested_xgb}
\alias{predict_nested_xgb}
\title{Get predicteds for a XGB fit from a nested.glmnetr() output object}
\usage{
predict_nested_xgb(object, xs, type = NULL, tuned = 1)
}
\arguments{
\item{object}{an output object from a nested.glmnetr() call}

\item{xs}{an new predictor matrix of same format as used in the nested.glmnetr() call}

\item{type}{NULL for the standard model, "feat" for the model including the 
realaxed lasso predicteds as a featuere,  "offs" for the model including the 
realaxed lasso predicteds as an offset.}

\item{tuned}{1 (default) for tuned xgb models, 0 for the simple untuned models}
}
\value{
predicteds from a XGB model fit in a nested.glmnetr() output object
}
\description{
Get predicteds for a XGB fit from a nested.glmnetr() output object
}
\seealso{
\code{\link{predict_ann_tab}} , \code{\link{predict_nested_rf}} , \code{\link{predict.nested.glmnetr}} , \code{\link{nested.glmnetr}}
}
