% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_cv_240322.R
\name{prednn_tl}
\alias{prednn_tl}
\title{predicted values from an ann_tab_cv output object based upon the model and its
lasso model used for generating an offset}
\usage{
prednn_tl(lassomod, nnmodel, datain, lasso = 1)
}
\arguments{
\item{lassomod}{a lasso model from a glmnetr() call used to genrte an offset}

\item{nnmodel}{a ann_tab_cv() output object for}

\item{datain}{new data}

\item{lasso}{1 if an offset is to be added as column 1 for calculations (default), 
0 to subset to the terms significant in a lasso model without adding the offset}
}
\value{
predictions from an nerual network model accounting from a lasso model
}
\description{
predicted values from an ann_tab_cv output object based upon the model and its
lasso model used for generating an offset
}
