% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObliqueRandomForest_240508.R
\name{print.orf_tune}
\alias{print.orf_tune}
\title{Print output from orf_tune() function}
\usage{
\method{print}{orf_tune}(x, ...)
}
\arguments{
\item{x}{output from an orf_tune() function}

\item{...}{optional pass through parameters to pass to print.orf()}
}
\value{
summary to console
}
\description{
Print output from orf_tune() function
}
\seealso{
\code{\link{summary.orf_tune}} , \code{\link{orf_tune}} , \code{\link{nested.glmnetr}}
}
