% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rederive_240508.R
\name{rederive_orf}
\alias{rederive_orf}
\title{Rederive Oblique Random Forest models not kept in nested.glmnetr() output}
\usage{
rederive_orf(object, xs, y_, event = NULL, type = NULL)
}
\arguments{
\item{object}{A nested.glmnetr() output object}

\item{xs}{Same xs used as input to ntested.glmnetr() for input object.}

\item{y_}{Same y_ used as input to ntested.glmnetr() for input object.}

\item{event}{Same event used as input to ntested.glmnetr() for input object.}

\item{type}{Same type used as input to ntested.glmnetr() for input object.}
}
\value{
an output like nested.glmnetr()$rf_tuned_fitX for X in c("", "F", "O")
}
\description{
Because the oblique random forest models sometimes take large 
amounts of storage one may decide to set keep=0 within the doorf list passed 
to nested.glmnetr().  This function allows the user to rederive the oblique 
random forest models without doing the search. Note, the oblique random 
forest fitting for survival data routine does not allow for (start,stop) 
times.
}
\seealso{
\code{\link{orf_tune}} , \code{\link{nested.glmnetr}}
}
