% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_240322.R
\name{rf_xbhat}
\alias{rf_xbhat}
\title{get XBeta from an rfsrc output object}
\usage{
rf_xbhat(rf_model, dframe, ofst = NULL, family, tol = 1e-05)
}
\arguments{
\item{rf_model}{an output from rfsrc()}

\item{dframe}{a data fraem of new data}

\item{ofst}{an offset on the XB scale}

\item{family}{one of c("cox", "binomial", "gaussian")}

\item{tol}{a small number, a lower bound to avoid division by 0}
}
\value{
the XBeta term from an rfsrc() output object
}
\description{
get XBeta from an rfsrc output object
}
