% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nested.glmnetr_240410.R
\name{roundperf}
\alias{roundperf}
\title{round elements of a summary.glmnetr() output}
\usage{
roundperf(summdf, digits = 3, do_ncv = 1)
}
\arguments{
\item{summdf}{a summary data frame from summary.nested.glmnetr() obtained using 
the option table=0}

\item{digits}{the minimum number of decimals to display the elements of the data
frame}

\item{do_ncv}{1 (default) if the summdf object is a summary for an analysis including
nested cross validation, 0 if only the full data models were fit.}
}
\value{
a data frame with same form as the input but with rounding for easier 
display
}
\description{
round elements of a summary.glmnetr() output
}
\seealso{
\code{\link{summary.nested.glmnetr}} , \code{\link{nested.glmnetr}}
}
