% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_240322.R
\name{rpart_xbhat}
\alias{rpart_xbhat}
\title{get XBeta from an rpart output object}
\usage{
rpart_xbhat(rpart_model, dframe, ofst = NULL, family, tol = 1e-05)
}
\arguments{
\item{rpart_model}{an rpart() output object}

\item{dframe}{A data frame as input}

\item{ofst}{An offset term}

\item{family}{family, one of "cox", "binomial" or "gaussian"}

\item{tol}{a cap at small or inverse of large numbers to avoid numeric overflow}
}
\value{
XBeta hat for processing to rpart_perf()
}
\description{
get XBeta from an rpart output object
}
