% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.glmnetr_250818.R
\name{summary.cv.glmnetr}
\alias{summary.cv.glmnetr}
\title{Output summary for elastic net models fit within a nested.glmnetr() output object.}
\usage{
\method{summary}{cv.glmnetr}(object, printg1 = "FALSE", orderall = FALSE, type = "lasso", ...)
}
\arguments{
\item{object}{a nested.glmnetr() output object.}

\item{printg1}{TRUE to also print out the fully penalized lasso beta, else FALSE to suppress.}

\item{orderall}{By default (orderall=FALSE) the order terms enter into the lasso model 
is given for the number of terms that enter in lasso minimizing loss model.  If 
orderall=TRUE then all terms that are included in any lasso fit are described.}

\item{type}{once of c("lasso", "elastic", "ridge") to select for summarizing,
with default of "lasso".}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
Coefficient estimates (beta)
}
\description{
Summarize the cross-validation informed model fit.  The fully penalized
(gamma=1) beta estimate will not be given by default but can too be output  
using printg1=TRUE.
}
\seealso{
\code{\link{predict.cv.glmnetr}} , \code{\link{nested.glmnetr}}
}
