% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.glmnetr_0_6_1_250501.R
\name{summary.cv.glmnetr_0_6_1}
\alias{summary.cv.glmnetr_0_6_1}
\title{Output summary for elastic net models fit within a nested.glmnetr() output object.}
\usage{
\method{summary}{cv.glmnetr_0_6_1}(
  object,
  type = NULL,
  printg1 = "FALSE",
  orderall = FALSE,
  betatol = 1e-14,
  ...
)
}
\arguments{
\item{object}{a nested.glmnetr() output object.}

\item{type}{one of c("lasso", "elastic", "ridge") to select for summarizing,
with default of "lasso".}

\item{printg1}{TRUE to also print out the fully penalized lasso beta, else FALSE to suppress.}

\item{orderall}{By default (orderall=FALSE) the order terms enter into the lasso model 
is given for the number of terms that enter in lasso minimizing loss model.  If 
orderall=TRUE then all terms that are included in any lasso fit are described.}

\item{betatol}{beta values less than betatol are set to 0 when they are 
approaching the rounding error of the machine architecture. Default is set 
to 1e-14.}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
Coefficient estimates (beta)
}
\description{
Summarize the cross-validation informed model fit.  The fully penalized
(gamma=1) beta estimate will not be given by default but can too be output  
using printg1=TRUE.
}
\seealso{
\code{\link{predict.cv.glmnetr}} , \code{\link{nested.glmnetr}}
}
