% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObliqueRandomForest_241202.R
\name{summary.orf_tune}
\alias{summary.orf_tune}
\title{Summarize output from rf_tune() function}
\usage{
\method{summary}{orf_tune}(object, ...)
}
\arguments{
\item{object}{output from an rf_tune() function}

\item{...}{optional pass through parameters to pass to summary.orsf()}
}
\value{
summary to console
}
\description{
Summarize output from rf_tune() function
}
\seealso{
\code{\link{rf_tune}} , \code{\link{nested.glmnetr}}
}
