% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_240322.R
\name{summary.rf_tune}
\alias{summary.rf_tune}
\title{Summarize output from rf_tune() function}
\usage{
\method{summary}{rf_tune}(object, ...)
}
\arguments{
\item{object}{output from an rf_tune() function}

\item{...}{optional pass through parameters to pass to summary.rfsrc()}
}
\value{
summary to console
}
\description{
Summarize output from rf_tune() function
}
\seealso{
\code{\link{predict_nested_rf}} , \code{\link{rf_tune}} , \code{\link{nested.glmnetr}}
}
