% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_cv_230509.R
\name{wtzero}
\alias{wtzero}
\title{Construct the weights for going from the observed data with an offset in column 1
to the first hidden layer}
\usage{
wtzero(tnsr, lasso = 0, rreturn = 1, trnspose = 0)
}
\arguments{
\item{tnsr}{an input tensor which is to be modified to mimic the linear term of a 
generalized linear model, e.g a Cox or logistic regression model}

\item{lasso}{1 if the first column is the linear estimate from a linear model, 
often a lasso model}

\item{rreturn}{1 (default) to return an R (numeric) vector, 0 to return a torch tensor}

\item{trnspose}{1 to transpose the matrix before returning, 0 to not.}
}
\value{
a weight matrix in tensor format
}
\description{
Construct the weights for going from the observed data with an offset in column 1
to the first hidden layer
}
