% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_240322.R
\name{xgb_xbhat}
\alias{xgb_xbhat}
\title{get XBeta from an XGB.train output object}
\usage{
xgb_xbhat(xgb_model, xs_, family, tol = 1e-05)
}
\arguments{
\item{xgb_model}{an output from xgb.train()}

\item{xs_}{and XGBoost data matrix}

\item{family}{one of c("cox", "binomial", "gaussian")}

\item{tol}{a small number, a lower bound to avoid division by 0}
}
\value{
the XBeta term from an XGB.train() output object
}
\description{
get XBeta from an XGB.train output object
}
