% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepreg_230218.R
\name{best.preds}
\alias{best.preds}
\title{Get the best models for the steps of a stepreg() fit}
\usage{
best.preds(modsum, risklist)
}
\arguments{
\item{modsum}{model summmary}

\item{risklist}{riskset list}
}
\value{
best predictors at each step of a stepwise regerssion
}
\description{
Get the best models for the steps of a stepreg() fit
}
