% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepreg_240422.R
\name{best.preds}
\alias{best.preds}
\title{Get the best models for the steps of a stepreg() fit}
\usage{
best.preds(modsum, risklist)
}
\arguments{
\item{modsum}{model summmary}

\item{risklist}{riskset list}
}
\value{
best predictors at each step of a stepwise regression
}
\description{
Get the best models for the steps of a stepreg() fit
}
\seealso{
\code{\link{stepreg}} , \code{\link{cv.stepreg}} , \code{\link{nested.glmnetr}}
}
