% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmnetr_230218.R
\name{difftime2}
\alias{difftime2}
\title{Output to console the elapsed and split times}
\usage{
difftime2(time_start = NULL, time_last = NULL)
}
\arguments{
\item{time_start}{beginning time for printing elapsed time}

\item{time_last}{last time for calculating split time}
}
\value{
Time of program invocation
}
\description{
Output to console the elapsed and split times
}
\examples{
time_start = difftime2()
time_last = difftime2(time_start)
time_last = difftime2(time_start,time_last)
time_last = difftime2(time_start,time_last)

}
