% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_241106.R
\name{get.foldid}
\alias{get.foldid}
\title{Get foldid's with branching for cox, binomial and gaussian models}
\usage{
get.foldid(y_, event, family, folds_n, stratified = 1)
}
\arguments{
\item{y_}{see help for cv.glmnetr() or nested.glmnetr()}

\item{event}{see help for cv.glmnetr() or nested.glmnetr()}

\item{family}{see help for cv.glmnetr() or nested.glmnetr()}

\item{folds_n}{see help for cv.glmnetr() or nested.glmnetr()}

\item{stratified}{see help for cv.glmnetr() or nested.glmnetr()}
}
\value{
A numeric vector with foldid's for use in a cross validation
}
\description{
Get foldid's with branching for cox, binomial and gaussian models
}
\seealso{
\code{\link{factor.foldid}} , \code{\link{nested.glmnetr}}
}
