% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.glmnetr_221210.R
\name{getlamgam}
\alias{getlamgam}
\title{get lam and gam}
\usage{
getlamgam(object, lam, gam, comment)
}
\arguments{
\item{object}{- glmnetr object as input}

\item{lam}{value for lam, may be NULL}

\item{gam}{value for gam, may be MULL}

\item{comment}{Default of TRUE to write to console information on lam and gam selected for output.
FALSE will suppress this write to console.}
}
\value{
numerical values for lam and gam for useage in plot and predict
}
\description{
get lam and gam
}
