% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.compare_0_5_1_241224.R,
%   R/nested.compare_0_5_2_241224.R, R/nested.compare_0_5_3_241224.R,
%   R/nested.compare_251212.R
\name{glmnetr.compcv}
\alias{glmnetr.compcv}
\title{A redirect to nested.compare}
\usage{
glmnetr.compcv(object, digits = 4, type = "devrat", pow = 1)

glmnetr.compcv(object, digits = 4, type = "devrat", pow = 1)

glmnetr.compcv(object, digits = 4, type = "devrat", pow = 1)

glmnetr.compcv(object, digits = 4, type = "devrat", pow = 1)
}
\arguments{
\item{object}{A nested.glmnetr output object.}

\item{digits}{digits for printing of z-scores, p-values, etc. with default of 4}

\item{type}{determines what type of nested cross validation performance measures are 
compared.  Possible values are "devrat" to compare the deviance ratios, i.e. the 
fractional reduction in deviance relative to the null model deviance, 
"agree" to compare agreement, "lincal" to compare the linear calibration 
slope coefficients, "intcal" to compare the linear calibration intercept 
coefficients, from the nested cross validation.}

\item{pow}{the power to which the average of correlations is to be raised.}
}
\value{
A printout to the R console.

A printout to the R console.

A printout to the R console.

A printout to the R console.
}
\description{
See nested.compare(), as glmnetr() is depricated

See nested.compare(), as glmnetr() is depricated

See nested.compare(), as glmnetr.compcv() is depricated

See nested.compare(), as glmnetr.compcv() is depricated
}
\seealso{
\code{\link{nested.compare}}

\code{\link{nested.compare}}
}
