% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nested.glmnetr_231230.R
\name{glmnetr.compcv0}
\alias{glmnetr.compcv0}
\title{A glmnetr specifc paired t-test}
\usage{
glmnetr.compcv0(a, b, digits = 4)
}
\arguments{
\item{a}{One term}

\item{b}{A second term}

\item{digits}{digits for printing of z-scores, p-values, etc. with default of 4}
}
\value{
A t-test
}
\description{
Perform a paired t-test as called from glmnetr.compcv().
}
