% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nested.glmnetr_230218.R
\name{glmnetr.compcv0}
\alias{glmnetr.compcv0}
\title{A glmnetr specifc paired t-test}
\usage{
glmnetr.compcv0(a, b)
}
\arguments{
\item{a}{One term}

\item{b}{A second term}
}
\value{
A t-test
}
\description{
Perform a paired t-test as called from glmnetr.compcv().
}
