% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepreg_240322.R
\name{preds_1}
\alias{preds_1}
\title{Get predictors form a stepwise regression model.}
\usage{
preds_1(modsumbest, k_, risklist, risklistl)
}
\arguments{
\item{modsumbest}{matrix with best predictors based upon number of model terms}

\item{k_}{Value for number of predictors in model}

\item{risklist}{Riskset list}

\item{risklistl}{Number of terms (length) in the riskset}
}
\value{
input to best.preds()
}
\description{
Get predictors form a stepwise regression model.
}
