% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepreg_240322.R
\name{summary.stepreg}
\alias{summary.stepreg}
\title{Briefly summarize steps in a stepreg() output object, i.e. a stepwise regression fit}
\usage{
\method{summary}{stepreg}(object, ...)
}
\arguments{
\item{object}{A stepreg() output object}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
Summarize a stepreg() object
}
\description{
Briefly summarize steps in a stepreg() output object, i.e. a stepwise regression fit
}
\seealso{
\code{\link{stepreg}} , \code{\link{cv.stepreg}} , \code{\link{nested.glmnetr}}
}
