% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepreg_221210.R
\name{summary.stepreg}
\alias{summary.stepreg}
\title{Title Give a brief summary of the steps in a stepwise regression fit}
\usage{
\method{summary}{stepreg}(object, ...)
}
\arguments{
\item{object}{A stepreg() output object}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
Summarize a stepreg object
}
\description{
Title Give a brief summary of the steps in a stepwise regression fit
}
